pyproject.toml
[project]
name = "onkyo-api"
version = "1.0.0"
description = "Modern RESTful API for network-enabled Onkyo receivers based on onkyo-eiscp library"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Onkyo API Contributors"}
]
keywords = ["onkyo", "eiscp", "api", "rest", "receiver"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Latest Flask and Connexion for OpenAPI 3.x support
    "flask>=3.1.0",
    "connexion[flask,swagger-ui,uvicorn]>=3.1.0",
    "werkzeug>=3.1.0",
    
    # Onkyo control library - using fork
    "onkyo-eiscp @ git+https://github.com/rajiteh/onkyo-eiscp.git",
    
    # Additional utilities
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "black>=24.10.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = ["main.py", "onkyo.py", "swagger.yml"]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "black>=24.10.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = []

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=. --cov-report=term-missing"
